Refactor OSC event handling
