+++ lua/tidal/highlighting/osc.lua
@@ -3,52 +3,18 @@ local OSC = {}
 local losc = require("losc.src.losc")
 local pluginLibUv = require("losc.src.losc.plugins.udp-libuv")
 
+local events = require("tidal.highlighting.events")
 local highlight = require("tidal.highlighting.highlights")
 local marker = require("tidal.highlighting.marker")
 
-OSC.messageBuffer = {}
-OSC.activeMessages = {}
+local messageBuffer = {}
+local activeMessages = {}
 
-local function buildMap(events)
-  local map = {}
-  for _, evt in ipairs(events) do
-    local key = evt.buf .. ":" .. evt.markerId
-    map[key] = evt
-  end
-  return map
-end
+OSC.timer = nil
 
-function OSC.diffEventLists(prevEvents, currentEvents)
-  local removed = {}
-  local added = {}
-  local active = {}
+local handleMessageCallback = nil
 
-  local prevMap = buildMap(prevEvents)
-  local currMap = buildMap(currentEvents)
-
-  -- Check removed and active
-  for key, prevEvt in pairs(prevMap) do
-    local currEvt = currMap[key]
-    if currEvt then
-      table.insert(active, prevEvt) -- still valid
-    else
-      table.insert(removed, prevEvt)
-    end
-  end
-
-  -- Check added
-  for key, currEvt in pairs(currMap) do
-    if not prevMap[key] then
-      table.insert(added, currEvt)
-    end
-  end
-
-  return {
-    removed = removed,
-    added = added,
-    active = active,
-  }
-end
+local uv = vim.uv
 
 local function startServer(host, port)
   local transport = pluginLibUv.new({ recvAddr = host, recvPort = port })
@@ -63,7 +29,7 @@ local function startServer(host, port)
       if marker.extMarks[eventId] and marker.extMarks[eventId][colStart] then
         local extmark = marker.extMarks[eventId][colStart]
         extmark.id = id
-        table.insert(OSC.messageBuffer, extmark)
+        table.insert(messageBuffer, extmark)
       else
         -- Drops -> Maybe count them?
         -- print(string.format("No extmark found at colStart=%s eventId=%s", tostring(colStart), tostring(eventId)))
@@ -82,12 +48,31 @@ local function startStyleServer(host, port)
     vim.schedule(function()
       local msg = data.message
       local id = msg[1]
-      local color = msg[2]
       highlight.addHl(id, color)
     end)
   end)
 
   osc:open()
+end
+
+local function handleMessages()
+  local diff = events.diffEventLists(activeMessages, messageBuffer)
+
+  for _, evt in ipairs(diff.added) do
+    highlight.addHighlight(evt.id, evt.buf, evt.markerId)
+  end
+
+  for _, evt in ipairs(diff.removed) do
+    highlight.removeHighlight(evt.buf, evt.markerId)
+  end
+
+  activeMessages = events.merge(diff.active, diff.added)
+
+  if handleMessageCallback then
+    handleMessageCallback(activeMessages)
+  end
+
+  messageBuffer = {}
 end
 
 function OSC.launch(highlightConf)
@@ -96,6 +81,22 @@ function OSC.launch(highlightConf)
 
   startServer(eventOsc.ip, eventOsc.port)
   startStyleServer(styleOsc.ip, styleOsc.port)
+end
+
+function OSC.setInterval(interval)
+  OSC.timer = uv.new_timer()
+  OSC.timer:start(interval, interval, function()
+    vim.schedule(handleMessages)
+  end)
+end
+
+function OSC.clearInterval()
+  if OSC.timer ~= nil then
+    OSC.timer:stop()
+    OSC.timer:close()
+    OSC.timer = nil
+  end
+  messageBuffer = {}
 end
 
 return OSC
